Add the update_records method using the updateMulti endpoint
Instead of using up the requests per 5 minutes quota, a new method that
will update several records for a domain at once would be a useful
addition.

So with that the method `update_records` came to be, it is using the
`updateMulti` endpoint, that is described in the api documentation.

It requires an array of records that contains the future values set to the
respective keys. This only applies to `name`,
`type` and `value` as these were the ones that seemed the most useful
ones.

The options hash is applied to all records in the update.

The records would look something like the following:

```rb
[
  {
    "id"=>"123",
    "name"=>"foo",
    "type"=>"A",
    "value"=>"127.0.0.1",
    "gtdLocation"=>"DEFAULT",
    "ttl"=>"300"
  },
  ....
]
```

method signature:

```rb
update_records('domain name', [array of records], {options})
```